+++ edbee-lib/edbee/views/textlayout.h
@@ -37,43 +37,42 @@ public:
 
     void buildLayout();
 
-    int toVirtualCursorPosition(int cursor) const;
-    int fromVirtualCursorPosition(int cursor) const;
+    size_t toVirtualCursorPosition(size_t cursor) const;
+    size_t fromVirtualCursorPosition(size_t cursor) const;
 
 
     void draw(QPainter *p, const QPointF &pos, const QVector<QTextLayout::FormatRange> &selections = QVector<QTextLayout::FormatRange>(), const QRectF &clip = QRectF()) const;
-    void drawCursor(QPainter *painter, const QPointF &position, int cursorPosition, int width) const;
+    void drawCursor(QPainter *painter, const QPointF &position, size_t cursorPosition, int width) const;
 
     void setFormats(const QVector<QTextLayout::FormatRange> &formats);
     void setText(const QString &string);
 
     void useSingleCharRanges();
     TextRangeSet* singleCharRanges() const;
-    void addSingleCharRange(int index, int length);
+    void addSingleCharRange(size_t index, size_t length);
 
 
     /// These calculations manipulate the cursor position. ..
     /// We assume the SOURCE character is 1... (TODO: store the orginal number of replace characters)
-    qreal cursorToX(int cursorPos, QTextLine::Edge edge = QTextLine::Leading) const;
-    int xToCursor(qreal x, QTextLine::CursorPosition cpos = QTextLine::CursorBetweenCharacters) const;
+    qreal cursorToX(size_t cursorPos, QTextLine::Edge edge = QTextLine::Leading) const;
+    size_t xToCursor(qreal x, QTextLine::CursorPosition cpos = QTextLine::CursorBetweenCharacters) const;
 
 protected:
     QTextLayout *qtextLayout_;
     TextDocument *textDocumentRef_;
-    TextRangeSet *singleCharRanges_;         ///< A list textRanges_ used by TextLayout. Every range in this list is treatet as a single character for cusor-movement etc
+    TextRangeSet *singleCharRanges_; ///< A list textRanges_ used by TextLayout. Every range in this list is treatet as a single character for cusor-movement etc
     QTextLine qtextLine_;
 };
 
 
 //=================================================
 
+
 class TextLayoutBuilder
 {
 public:
     TextLayoutBuilder(TextLayout* textLayout, QString & baseString, QVector<QTextLayout::FormatRange> & baseFormatRanges);
-
-    void replace(int index, int length, const QString replacement, QTextCharFormat format);
-
+    void replace(size_t index, size_t length, const QString replacement, QTextCharFormat format);
 
 protected:
     TextLayout* textLayoutRef_;                                ///< A reference to the textlayout
